+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.1] - 2017-12-11
+### Changed
+ - Loading `Twig` classes with psr-4 mode;
+ - Select `option` tag don't render `value` if it is empty.
+
 ## [2.0.0] - 2017-12-11
 ### Added
  - Documentation of package;